refactor: :recycle: refactor tests paths
